+++ demo/uni-app-vue3-vite/src/pages/index/index.vue
@@ -1,5 +1,7 @@
 <template>
   <view class="content">
+    <view class="test">test</view>
+    <view :key="i" class="h-[20px] w-[20px]" :class="cardsColor[i - 1]" v-for="i in 5"></view>
     <!-- <view class="w-2 h-2 bg-[#123456]"></view>
     <view class="w-2 h-2 bg-blue-500/50"></view> -->
     <view class="flex items-center justify-center w-screen h-screen">
@@ -27,8 +29,7 @@
       </view>
       <view class="w-32 py-2 rounded-md font-semibold text-white bg-pink-500 ring-4 ring-pink-300"> Default </view>
     </view>
-    <view class="test">test</view>
-    <view :key="i" class="h-[20px] w-[20px]" :class="cardsColor[i - 1]" v-for="i in 5"></view>
+
   </view>
 </template>
 
@@ -36,8 +37,9 @@
 import { ref, reactive } from "vue";
 const title = ref("测试标题");
 const flag = ref(true);
+import { replaceJs } from '../../../../..'
 const cardsColor = reactive([
-  'bg-[#4268EA] shadow-indigo-100',
+  replaceJs('bg-[#4268EA] shadow-indigo-100'),
   'bg-cyan-500 shadow-blue-100',
   'bg-green-500 shadow-green-100',
   'bg-cyan-500 shadow-cyan-100',
