+++ rollup.config.js
@@ -6,34 +6,57 @@ import pkg from './package.json'
 // const isProd = process.env.NODE_ENV === 'production'
 const isDev = process.env.NODE_ENV === 'development'
 
-/** @type {import('rollup').RollupOptions} */
-const config = {
-  input: 'src/index.ts',
-  output: [
-    {
-      file: pkg.main,
-      format: 'cjs',
-      sourcemap: isDev
-      // exports: 'auto'
-    }
-    // { format: 'esm', file: pkg.module, sourcemap: isDev }
-  ],
+/** @type {Array<import('rollup').RollupOptions> } */
+const config = [
+  {
+    input: 'src/index.ts',
+    output: [
+      {
+        file: pkg.main,
+        format: 'cjs',
+        sourcemap: isDev
+        // exports: 'auto'
+      }
+      // { format: 'esm', file: pkg.module, sourcemap: isDev }
+    ],
 
-  plugins: [
-    nodeResolve({
-      preferBuiltins: true
-    }),
-    commonjs(),
-    typescript({ tsconfig: './tsconfig.build.json', sourceMap: isDev, declaration: false })
-    // isDev
-    //   ? undefined
-    //   : terser({
-    //     format: {
-    //       comments: false,
-    //     }
-    //   })
-  ],
-  external: [...(pkg.dependencies ? Object.keys(pkg.dependencies) : [])]
-}
+    plugins: [
+      nodeResolve({
+        preferBuiltins: true
+      }),
+      commonjs(),
+      typescript({ tsconfig: './tsconfig.build.json', sourceMap: isDev, declaration: false })
+      // isDev
+      //   ? undefined
+      //   : terser({
+      //     format: {
+      //       comments: false,
+      //     }
+      //   })
+    ],
+    external: [...(pkg.dependencies ? Object.keys(pkg.dependencies) : [])]
+  },
+  {
+    input: 'src/replace.ts',
+    output: [
+      {
+        file: 'dist/replace.js',
+        format: 'esm',
+        sourcemap: isDev
+        // exports: 'auto'
+      }
+      // { format: 'esm', file: pkg.module, sourcemap: isDev }
+    ],
+
+    plugins: [
+      nodeResolve({
+        preferBuiltins: true
+      }),
+      commonjs(),
+      typescript({ tsconfig: './tsconfig.build.json', sourceMap: isDev, declaration: false })
+    ],
+    external: [...(pkg.dependencies ? Object.keys(pkg.dependencies) : [])]
+  }
+]
 
 export default config
