+++ src/replace.ts
@@ -1,0 +1,4 @@
+import { replaceWxml as replaceJs } from './wxml/shared'
+import { cssSelectorReplacer as replaceCss } from './postcss/shared'
+
+export { replaceJs, replaceCss }
